// frontend/bindings/github.com/telesma-app/kit/model/models.ts
export enum InteractionKind {
    /**
     * The Go zero value for the underlying type of the enum.
     */
    $zero = "",

    InteractionKindPIN = "pin",
    InteractionKindUserVerification = "user-verification",
    InteractionKindTouch = "touch",
};